+++ django-server/server/admin/apps.py
@@ -1,0 +1,8 @@
+from django.apps import AppConfig
+
+
+class AdminConfig(AppConfig):
+    name = 'admin'
+    label= 'myadmin'
+    default_site = 'admin.MyAdminSite'
+
