+++ django-server/server/admin/templates/stats.html
