+++ django-server/server/server/urls.py
@@ -13,7 +13,7 @@ Including another URLconf
     1. Import the include() function: from django.urls import include, path
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
-from api.admin import admin_site
+from admin.admin import admin_site
 from django.contrib import admin
 from django.urls import include, path
 from rest_framework import permissions
